requirements.txt
azure-cognitiveservices-vision-customvision>=3.1.0
botbuilder-integration-aiohttp>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-ai>=4.14.0
